requirements.txt
astapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0